+++ src/dgloss/print.py
@@ -5,17 +5,14 @@ import pastel
 import dgloss
 
 
-class Formatter:
+class Printer:
     def __init__(self):
         if dgloss.disable_ansi:
             pastel.with_colors(False)
         else:
             pastel.with_colors(True)
 
-    def print(self, msg, stdout=True):
-        msg = pastel.colorize(msg)
+    def print(self, msg, file=sys.stdout):
+        msg = pastel.colorize(str(msg))
         msg = f"{msg}\n"
-        if stdout:
-            sys.stdout.write(msg)
-        else:
-            sys.stderr.write(msg)
+        file.write(msg)